+++ DEDIKASI_RPL_WebProject/app/Http/Controllers/ProfileMentorController.php
@@ -39,8 +39,8 @@ class ProfileMentorController extends Controller
         if ($request->password != '') {
 
             $request->validate([
-                'password' => 'required|min:8',
-                'password_confirm' => 'required|min:8|same:password' ,
+                'password' => 'required',
+                'password_confirm' => 'required|same:password' ,
             ]);
             
             $data_profile_mentor->password = Hash::make($request->password_confirm);
@@ -63,10 +63,14 @@ class ProfileMentorController extends Controller
                 'photo' => 'image|mimes:jpg,png,jpeg',
             ]);
 
+            if ($data_profil_mentor->foto_mentor != 'default_image.jpg') {
+                
+                unlink(public_path('assets/img/profiles/'.$data_profil_mentor->foto_mentor));
+            }
+
             $time = time();
-            unlink(public_path('assets/img/profiles/'.$data_profil_mentor->foto_mentor));
             $ext = $request->file('photo')->extension();
-            $photo_name = $data_profil_mentor->email.$time.$ext;
+            $photo_name = $data_profil_mentor->email.$time.'.'.$ext;
             $request->file('photo')->move(public_path('assets/img/profiles/'), $photo_name);
             $data_profil_mentor->foto_mentor = $photo_name;
         }
